Split up some single_match UI tests
This moves only the single_match tests over to the new file.
